# idoxxy-openapi.pretty.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Trwaly Nosnik API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://api.idoxxy.com"
        }
    ],
    "security": [
        {
            "main": []
        },
        {
            "AccountApiKeyAuth": []
        }
    ],
    "paths": {
...
        "/groups": {
            "post": {
                "tags": [
                    "group-controller"
                ],
                "operationId": "createGroup",
                "parameters": [
                    {
                        "name": "groupName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "format": "uuid"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Problem"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupToListWithCustomers"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "main": []
                    },
                    {
                        "AccountApiKeyAuth": []
                    }
                ]
            }
        },
...
        "/documents": {
            "post": {
                "tags": [
                    "document-controller"
                ],
                "operationId": "createDocument",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "request"
                                ],
                                "type": "object",
                                "properties": {
                                    "request": {
                                        "$ref": "#/components/schemas/CreateDocumentRequest"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Problem"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "main": []
                    },
                    {
                        "AccountApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Problem": {
                "type": "object",
                "properties": {
                    "detail": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/StatusType"
                    },
                    "instance": {
                        "type": "string",
                        "format": "uri"
                    },
                    "type": {
                        "type": "string",
                        "format": "uri"
                    },
                    "parameters": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    }
                }
            },
            "StatusType": {
                "type": "object",
                "properties": {
                    "reasonPhrase": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "CustomerToList": {
                "required": [
                    "createdAt",
                    "email",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    }
                }
            },
            "GroupToListWithCustomers": {
                "required": [
                    "createdAt",
                    "customers",
                    "deferred",
                    "groupName",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "groupName": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "customers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CustomerToList"
                        }
                    },
                    "deferred": {
                        "type": "boolean"
                    }
                }
            },
...
            "Document": {
                "required": [
                    "createdAt",
                    "documentName",
                    "documentType",
                    "id",
                    "userId",
                    "versions"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "documentName": {
                        "type": "string"
                    },
                    "versions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DocumentVersion"
                        }
                    },
                    "documentType": {
                        "type": "string",
                        "enum": [
                            "GROUP",
                            "INDIVIDUAL"
                        ]
                    },
                    "assignedIds": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "currentVersion": {
                        "$ref": "#/components/schemas/DocumentVersion"
                    }
                }
            },
            "DocumentVersion": {
                "required": [
                    "createdAt",
                    "documentInfo",
                    "fileId",
                    "id",
                    "scheduled",
                    "validFrom",
                    "versionId",
                    "versionStatus"
                ],
                "type": "object",
                "properties": {
                    "documentInfo": {
                        "$ref": "#/components/schemas/Document"
                    },
                    "fileId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "versionId": {
                        "type": "string"
                    },
                    "validFrom": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "validTo": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "scheduled": {
                        "type": "boolean"
                    },
                    "versionStatus": {
                        "type": "string",
                        "enum": [
                            "Active",
                            "Inactive",
                            "Archived"
                        ]
                    }
                }
            },
...
            "CreateDocumentRequest": {
                "required": [
                    "documentType",
                    "fileName",
                    "recipients",
                    "validFrom"
                ],
                "type": "object",
                "properties": {
                    "fileName": {
                        "type": "string"
                    },
                    "documentType": {
                        "type": "string",
                        "enum": [
                            "GROUP",
                            "INDIVIDUAL"
                        ]
                    },
                    "recipients": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "validFrom": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "validTo": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "main": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token"
                    },
                    "clientCredentials": {
                        "tokenUrl": "/oauth/token"
                    }
                }
            },
            "AccountApiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    }
}